pyproject.toml
[project]
name = "oneroof"
version = "2.0.0"
description = "Base-, Variant-, and Consensus-calling under One Proverbial Roof."
authors = [
  { name = "Nicholas R. Minor", email = "nrminor@wisc.edu" },
  { name = "Annabelle Kalweit", email = "akalweit@wisc.edu" },
  { name = "Andrew J. Lail", email = "ajlail@wisc.edu" },
  { name = "William K. Gardner", email = "wkgardner@wisc.edu" },
]
requires-python = ">= 3.10"
keywords = [
  "science",
  "biology",
  "research",
  "bioinformatics",
  "sequencing",
  "genomics",
  "virus",
  "SARS-CoV-2",
  "H5N1",
  "influenza",
  "haplotyping",
]
classifiers = [
  "Development Status :: 3 - Alpha",

  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Bio-Informatics",

  "License :: OSI Approved :: MIT License",

  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Rust",
  "Programming Language :: Other",          # Nextflow (Groovy-based DSL)
]
dependencies = [
  "biopython>=1.84",
  "pysam>=0.22.1",
  "scipy>=1.14.1",
  "paramiko>=3.5.0",
  "plotnine>=0.14.4",
  "loguru>=0.7.0",
  "edlib>=1.3.9.post1",
  "lxml>=5.3.0",
  "pyyaml",
  "pytz>=2025.1",
  "pydantic>=2.9.2,<3",
  "polars-lts-cpu[calamine,numpy,openpyxl,pandas,pyarrow,xlsx2csv,xlsxwriter]>=1.22.0",
  "multiqc>=1.33",
  "slack-sdk>=3.39.0",
  "duckdb>=1.2.0",
  "rich>=14.0.0",
  "typer>=0.20.0",
  "altair>=6.0.0",
  "vl-convert-python>=1.8.0",
]

[tool.pixi.workspace]
channels = ["conda-forge", "bioconda", "anaconda"]
platforms = ["osx-64", "linux-64", "osx-arm64"]

[tool.pixi.activation]
env = { RUSTFLAGS = "-C target-cpu=native" }

[tool.pixi.dependencies]
nextflow = ">=24.4.2,<24.5"
bbmap = ">=39.1,<40"
fastp = ">=0.24.0,<0.25"
vsearch = ">=2.28.1,<2.29"
seqkit = ">=2.8.2,<2.9"
bedtools = ">=2.31.1,<3"
samtools = ">=1.5,<2"
minimap2 = ">=2.28,<3"
bcftools = ">=1.16,<2"
ivar = ">=1.4.4,<2"
snpeff = ">=5.2,<6"
snpsift = ">=5.2,<6"
fastqc = ">=0.12.1,<0.13"
nextclade = ">=3.18.1,<4"
chopper = ">=0.9.0,<0.10"
rasusa = ">=2.0.0,<2.1"
deacon = ">=0.2.0,<0.3"
devider = ">=0.0.1,<0.0.2"
sylph = ">=0.8.1,<0.9"
sylph-tax = ">=1.2.0,<2"
rust-script = ">=0.36.0,<0.37"

[tool.pixi.target.osx-64.dependencies]
fq = { version = ">=0.12,<0.13", channel = "bioconda" }

[tool.pixi.target.linux-64.dependencies]
fq = { version = ">=0.12,<0.13", channel = "bioconda" }

[dependency-groups]
dev = [
  "basedpyright>=1.21.0",
  "ipykernel>=6.29.5",
  "ipython>=8.12.3",
  "jupyter>=0.6.4",
  # "marimo[sql]>=0.18.4",
  "pip>=24.2",
  "pytest>=8.3.3",
  "quarto>=0.1.0",
  "ruff>=0.9.6",
  "tox>=4.19",
  "tox-uv>=1.15",
  "ty>=0.0.3",
]

[tool.uv]
compile-bytecode = true

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = "lib"
module-name = ["oneroof_cli", "reporting"]

[project.scripts]
"oneroof" = "oneroof_cli:main"

[tool.pixi.pypi-dependencies]
oneroof = { path = ".", editable = true }

[tool.basedpyright]
venvPath = "."
venv = ".venv"
typeCheckingMode = "basic"

[tool.pytest.ini_options]
pythonpath = ["bin", "lib"]
testpaths = ["bin", "lib"]

[tool.tox]
requires = ["tox>=4.19", "tox-uv>=1.15"]
env_list = ["py310", "py311", "py312", "py313", "lint"]

[tool.tox.env_run_base]
description = "Run tests under {base_python}"
runner = "uv-venv-lock-runner"
dependency_groups = ["dev"]
commands = [["pytest"]]

[tool.tox.env.lint]
description = "Run ruff linter"
skip_install = true
commands = [["ruff", "check", "."]]
